Conflicts:
	imi/src/test/java/org/motechproject/nms/imi/it/TargetFileServiceBundleIT.java
	pom.xml

+++ mobile-academy/src/main/java/org/motechproject/nms/mobileacademy/service/MobileAcademyService.java
@@ -24,7 +24,7 @@ public interface MobileAcademyService {
      * Gets the course modification date as an epoch representation. This defaults to MobileAcademyCourse name
      * @return int representation (epoch) of modified course date
      */
-    int getCourseVersion();
+    long getCourseVersion();
 
     /**
      * Get the bookmark for a caller
